Validation Library Distribution V2
The files uploaded in this commit are those for the second version of
Mia’s Validation Library.

This version of the library is much less bloated and much more
efficiently coded compared to the last version. Method specific
variables have been removed in favour of instantiating different
objects from the Validation class. Please read/run the documentation
files (specifically, the README.txt file and the form.php file) in
order for detailed instructions on how to use the library. Please don’t
hesitate to contact me through email or through GitHub if you have any
questions or comments!